+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `machine` field to `StepResources`. You can set this to "local" when using the `BeakerExecutor` to force it to run the step locally.
 - Added `--ext-var` argument to `tango run` for setting JSONNET external variables
   when loading the experiment config.
+- Added `@step()` decorator to create `Step` classes from functions.
 
 ### Removed