+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug where importing certain libraries (like torchmetrics) would mess with our exception handling because they set `sys.excepthook` for some reason. Now we always reset `sys.excepthook` after importing.
 - The type hints for the flax trainer suggested that the training split is optional when in fact it's mandatory.
 - Made `BeakerWorkspace` / `BeakerStepLock` more robust when a job is preempted.
+- Minor performance improvements for the Beaker executor and workspace.
 
 
 ## [v1.0.2](https://github.com/allenai/tango/releases/tag/v1.0.2) - 2022-11-14